requirements.txt
feedparser>=6.0.10
matplotlib
networkx
newspaper3k>=0.2.8
pandas
urllib3>=1.26.13
python_dateutil>=2.8.2
pytz>=2022.6
requests>=2.28.1
tqdm>=4.64.1
underthesea>=6.0.0
torch